Update: Define pin to realize pin-level ATE test infrastructure.(BETA-1.0)

+++ C++/include/Timing.h
@@ -4,19 +4,19 @@
 #include <string>
 
 struct SingleEdgeTiming {
-    uint32_t edge = 1;
-    int32_t base = 0;
+    uint64_t edge = 1;
+    int64_t base = 0;
 };
 
 struct TwoEdgeTiming {
-    uint32_t edge_1 = 1;
-    uint32_t edge_2 = 3;
-    int32_t base = 0;
+    uint64_t edge_1 = 1;
+    uint64_t edge_2 = 3;
+    int64_t base = 0;
 };
 
 struct TimingSet {
     std::string name = "TS0";
-    uint32_t prd = 10;
+    uint64_t prd = 10;
     SingleEdgeTiming nrz{1, 0};
     TwoEdgeTiming rz{1, 3, 0};
     TwoEdgeTiming rzz{2, 7, 0};
